(BSR)[PRO] fix: bookings download returning 400 error
