+++ pro/src/pages/Bookings/__specs__/BookingsRecap.spec.tsx
@@ -10,6 +10,7 @@ import {
 import { ApiError } from 'apiClient/v2'
 import { ApiRequestOptions } from 'apiClient/v2/core/ApiRequestOptions'
 import { ApiResult } from 'apiClient/v2/core/ApiResult'
+import { FETCH_ERROR_MESSAGE } from 'app/App/App'
 import Notification from 'components/Notification/Notification'
 import { DEFAULT_PRE_FILTERS } from 'core/Bookings/constants'
 import {
@@ -311,11 +312,11 @@ describe('components | BookingsRecap | Pro user', () => {
       1,
       null,
       null,
-      'all',
+      null,
       DEFAULT_PRE_FILTERS.bookingStatusFilter,
       DEFAULT_PRE_FILTERS.bookingBeginningDate,
       DEFAULT_PRE_FILTERS.bookingEndingDate,
-      'all'
+      null
     )
   })
 
@@ -335,10 +336,7 @@ describe('components | BookingsRecap | Pro user', () => {
     await userEvent.click(downloadSubButton)
 
     expect(
-      await screen.findByText(
-        'Une erreur s’est produite. Veuillez réessayer ultérieurement',
-        { exact: false }
-      )
+      await screen.findByText(FETCH_ERROR_MESSAGE, { exact: false })
     ).toBeInTheDocument()
   })
