+++ pro/src/pages/CollectiveBookings/adapters/index.ts
@@ -1,4 +1,2 @@
 export { getFilteredCollectiveBookingsRecapAdapter } from './getFilteredCollectiveBookingsRecapAdapter'
-export { getCollectiveBookingsCSVFileAdapter } from './getCollectiveBookingsCSVFileAdapter'
-export { getCollectiveBookingsXLSFileAdapter } from './getCollectiveBookingsXLSFileAdapter'
 export { getUserHasCollectiveBookingsAdapter } from './getUserHasCollectiveBookingsAdapter'
