+++ pro/src/screens/Bookings/PreFilters/PreFilters.tsx
@@ -2,6 +2,7 @@ import classNames from 'classnames'
 import isEqual from 'lodash.isequal'
 import React, { useCallback, useEffect, useState } from 'react'
 
+import { FETCH_ERROR_MESSAGE } from 'app/App/App'
 import FormLayout from 'components/FormLayout/FormLayout'
 import MultiDownloadButtonsModal from 'components/MultiDownloadButtonsModal/MultiDownloadButtonsModal'
 import { DEFAULT_PRE_FILTERS } from 'core/Bookings/constants'
@@ -10,12 +11,10 @@ import { Audience } from 'core/shared'
 import useAnalytics from 'hooks/useAnalytics'
 import useNotification from 'hooks/useNotification'
 import fullRefreshIcon from 'icons/full-refresh.svg'
-import { getBookingsCSVFileAdapter } from 'pages/Bookings/adapters/getBookingsCSVFileAdapter'
-import { getBookingsXLSFileAdapter } from 'pages/Bookings/adapters/getBookingsXLSFileAdapter'
-import {
-  getCollectiveBookingsCSVFileAdapter,
-  getCollectiveBookingsXLSFileAdapter,
-} from 'pages/CollectiveBookings/adapters'
+import { downloadBookingsCSVFile } from 'pages/Bookings/downloadBookingsCSVFile'
+import { downloadBookingsXLSFile } from 'pages/Bookings/downloadBookingsXLSFile'
+import { downloadCollectiveBookingsCSVFile } from 'pages/CollectiveBookings/downloadCollectiveBookingsCSVFile'
+import { downloadCollectiveBookingsXLSFile } from 'pages/CollectiveBookings/downloadCollectiveBookingsXLSFile'
 import { Button } from 'ui-kit/Button/Button'
 import { ButtonVariant } from 'ui-kit/Button/types'
 import { isDateValid } from 'utils/date'
@@ -139,28 +138,23 @@ export const PreFilters = ({
     async (filters: PreFiltersParams, type: string) => {
       setIsDownloadingCSV(true)
 
-      if (audience === Audience.INDIVIDUAL) {
-        /* istanbul ignore next: DEBT to fix */
-        const { isOk, message } =
-          type === 'CSV'
-            ? await getBookingsCSVFileAdapter(filters)
-            : await getBookingsXLSFileAdapter(filters)
-
-        /* istanbul ignore next: DEBT to fix */
-        if (!isOk) {
-          notify.error(message)
+      try {
+        if (audience === Audience.INDIVIDUAL) {
+          /* istanbul ignore next: DEBT to fix */
+          if (type === 'CSV') {
+            await downloadBookingsCSVFile(filters)
+          } else {
+            await downloadBookingsXLSFile(filters)
+          }
+        } else {
+          if (type === 'CSV') {
+            await downloadCollectiveBookingsCSVFile(filters)
+          } else {
+            await downloadCollectiveBookingsXLSFile(filters)
+          }
         }
-      } else {
-        /* istanbul ignore next: DEBT to fix */
-        const { isOk, message } =
-          type === 'CSV'
-            ? await getCollectiveBookingsCSVFileAdapter(filters)
-            : await getCollectiveBookingsXLSFileAdapter(filters)
-
-        /* istanbul ignore next: DEBT to fix */
-        if (!isOk) {
-          notify.error(message)
-        }
+      } catch {
+        notify.error(FETCH_ERROR_MESSAGE)
       }
 
       setIsDownloadingCSV(false)
